refactor : #275 도메인 서비스 합병
