+++ payment/src/main/java/com/goti/payment/service/domain/ResaleEscrowService.java
@@ -4,8 +4,12 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import com.goti.payment.domain.entity.payment.EscrowAccountEntity;
+import com.goti.payment.dto.request.ResalePaymentRequest;
 
 public interface ResaleEscrowService {
+
+	List<EscrowAccountEntity> createEscrows(ResalePaymentRequest request);
+
 	List<EscrowAccountEntity> filterHoldings(List<EscrowAccountEntity> escrows);
 
 	void settle(List<EscrowAccountEntity> escrows, LocalDateTime releaseTime);
