--- payment/src/main/java/com/goti/payment/service/domain/ResaleOrderPaymentService.java
@@ -1,10 +1,0 @@
-package com.goti.payment.service.domain;
-
-import java.util.List;
-
-import com.goti.payment.domain.entity.payment.EscrowAccountEntity;
-import com.goti.payment.dto.request.ResalePaymentRequest;
-
-public interface ResaleOrderPaymentService {
-	List<EscrowAccountEntity> createEscrows(ResalePaymentRequest request);
-}
